# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CRM — Реєстр порушень та шахрайства
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Record {
  id: number;
  /**
     * Дата
     * @nullable
     */
  date?: string | null;
  /**
     * Відносно мережі/ГО
     * @nullable
     */
  networkOrg?: string | null;
  /**
     * по Факту
     * @nullable
     */
  byFact?: string | null;
  /**
     * Підстава
     * @nullable
     */
  basis?: string | null;
  /**
     * Ознаки шахрайства
     * @nullable
     */
  fraudSigns?: string | null;
  /**
     * Внутрішнє шахрайство
     * @nullable
     */
  internalFraud?: string | null;
  /**
     * Виявлено порушень
     * @nullable
     */
  violationsFound?: string | null;
  /**
     * Завдано збитків
     * @nullable
     */
  damagesCaused?: string | null;
  /**
     * Притягнуто до відповідальності/прийняті заходи
     * @nullable
     */
  measuresТaken?: string | null;
  /**
     * Передано в ОВС
     * @nullable
     */
  transferredToPolice?: string | null;
  /**
     * Результат ОВС
     * @nullable
     */
  policeResult?: string | null;
  /**
     * Заявник / Потерпілий
     * @nullable
     */
  applicantVictim?: string | null;
  /**
     * Стан контролю досудового розслідування
     * @nullable
     */
  investigationStatus?: string | null;
  /**
     * Додаткові нотатки
     * @nullable
     */
  notes?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface RecordInput {
  /** @nullable */
  date?: string | null;
  /** @nullable */
  networkOrg?: string | null;
  /** @nullable */
  byFact?: string | null;
  /** @nullable */
  basis?: string | null;
  /** @nullable */
  fraudSigns?: string | null;
  /** @nullable */
  internalFraud?: string | null;
  /** @nullable */
  violationsFound?: string | null;
  /** @nullable */
  damagesCaused?: string | null;
  /** @nullable */
  measuresТaken?: string | null;
  /** @nullable */
  transferredToPolice?: string | null;
  /** @nullable */
  policeResult?: string | null;
  /** @nullable */
  applicantVictim?: string | null;
  /** @nullable */
  investigationStatus?: string | null;
  /** @nullable */
  notes?: string | null;
}

export interface RecordUpdate {
  /** @nullable */
  date?: string | null;
  /** @nullable */
  networkOrg?: string | null;
  /** @nullable */
  byFact?: string | null;
  /** @nullable */
  basis?: string | null;
  /** @nullable */
  fraudSigns?: string | null;
  /** @nullable */
  internalFraud?: string | null;
  /** @nullable */
  violationsFound?: string | null;
  /** @nullable */
  damagesCaused?: string | null;
  /** @nullable */
  measuresТaken?: string | null;
  /** @nullable */
  transferredToPolice?: string | null;
  /** @nullable */
  policeResult?: string | null;
  /** @nullable */
  applicantVictim?: string | null;
  /** @nullable */
  investigationStatus?: string | null;
  /** @nullable */
  notes?: string | null;
}

export interface Stats {
  /** Всього записів */
  total: number;
  /** З ознаками шахрайства */
  withFraudSigns: number;
  /** Внутрішнє шахрайство */
  internalFraud: number;
  /** Передано в ОВС */
  transferredToPolice: number;
  /** Загальна сума збитків */
  totalDamages: string;
}

export interface ImportResult {
  /** Успішно імпортовано */
  imported: number;
  /** Пропущено (дублікати) */
  skipped: number;
  /** Помилки */
  errors: number;
  messages?: string[];
}

export type ListRecordsParams = {
/**
 * Пошук по всіх полях
 */
search?: string;
/**
 * Тільки записи з ознаками шахрайства (true/false)
 */
fraudOnly?: string;
/**
 * Тільки внутрішнє шахрайство (true/false)
 */
internalFraud?: string;
/**
 * Дата від (YYYY-MM-DD)
 */
dateFrom?: string;
/**
 * Дата до (YYYY-MM-DD)
 */
dateTo?: string;
};