# lib/api-zod/src/generated/types/importResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CRM — Реєстр порушень та шахрайства
 * OpenAPI spec version: 0.1.0
 */

export interface ImportResult {
  /** Успішно імпортовано */
  imported: number;
  /** Пропущено (дублікати) */
  skipped: number;
  /** Помилки */
  errors: number;
  messages?: string[];
}