# lib/api-zod/src/generated/types/record.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CRM — Реєстр порушень та шахрайства
 * OpenAPI spec version: 0.1.0
 */

export interface Record {
  id: number;
  /**
     * Дата
     * @nullable
     */
  date?: string | null;
  /**
     * Відносно мережі/ГО
     * @nullable
     */
  networkOrg?: string | null;
  /**
     * по Факту
     * @nullable
     */
  byFact?: string | null;
  /**
     * Підстава
     * @nullable
     */
  basis?: string | null;
  /**
     * Ознаки шахрайства
     * @nullable
     */
  fraudSigns?: string | null;
  /**
     * Внутрішнє шахрайство
     * @nullable
     */
  internalFraud?: string | null;
  /**
     * Виявлено порушень
     * @nullable
     */
  violationsFound?: string | null;
  /**
     * Завдано збитків
     * @nullable
     */
  damagesCaused?: string | null;
  /**
     * Притягнуто до відповідальності/прийняті заходи
     * @nullable
     */
  measuresТaken?: string | null;
  /**
     * Передано в ОВС
     * @nullable
     */
  transferredToPolice?: string | null;
  /**
     * Результат ОВС
     * @nullable
     */
  policeResult?: string | null;
  /**
     * Заявник / Потерпілий
     * @nullable
     */
  applicantVictim?: string | null;
  /**
     * Стан контролю досудового розслідування
     * @nullable
     */
  investigationStatus?: string | null;
  /**
     * Додаткові нотатки
     * @nullable
     */
  notes?: string | null;
  createdAt: string;
  updatedAt: string;
}