# lib/api-zod/src/generated/types/recordInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CRM — Реєстр порушень та шахрайства
 * OpenAPI spec version: 0.1.0
 */

export interface RecordInput {
  /** @nullable */
  date?: string | null;
  /** @nullable */
  networkOrg?: string | null;
  /** @nullable */
  byFact?: string | null;
  /** @nullable */
  basis?: string | null;
  /** @nullable */
  fraudSigns?: string | null;
  /** @nullable */
  internalFraud?: string | null;
  /** @nullable */
  violationsFound?: string | null;
  /** @nullable */
  damagesCaused?: string | null;
  /** @nullable */
  measuresТaken?: string | null;
  /** @nullable */
  transferredToPolice?: string | null;
  /** @nullable */
  policeResult?: string | null;
  /** @nullable */
  applicantVictim?: string | null;
  /** @nullable */
  investigationStatus?: string | null;
  /** @nullable */
  notes?: string | null;
}