# lib/api-zod/src/generated/types/stats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CRM — Реєстр порушень та шахрайства
 * OpenAPI spec version: 0.1.0
 */

export interface Stats {
  /** Всього записів */
  total: number;
  /** З ознаками шахрайства */
  withFraudSigns: number;
  /** Внутрішнє шахрайство */
  internalFraud: number;
  /** Передано в ОВС */
  transferredToPolice: number;
  /** Загальна сума збитків */
  totalDamages: string;
}